requirements.txt
tensorflow==2.13.0
numpy==1.23.5
Pillow==9.3.0
matplotlib==3.6.2
scipy==1.10.1